Added unit tests and updated documentation for Sequence.isEmpty.
